refactor(core): share OpenTelemetry noop between api and pure entries
Addresses review feedback on #9281: extracts the noop OpenTelemetry API factories and singleton from packages/core/src/instrumentation/api.ts into a new packages/core/src/instrumentation/noop.ts, and has both ./api and ./pure.index import from it. Prevents drift between the real (dynamic-import-with-catch) and pure (noop-only) entries.
